refacto : move beat manifest
