+++ frontend/src/app/infrastructure/adapters/beat-source/beat-adapter.service.ts
@@ -6,7 +6,7 @@ import { CompactBeatMapper } from "./compact-beat.mapper";
 import { jsonFileReaderToken } from "../../injection-tokens/json-file-reader.token";
 import { Array, Effect, Option } from "effect";
 import { Track } from "src/app/domain/track";
-import {BeatMetadata, BEATS_MANIFEST} from "../../../ui/services/sequencer/beats-manifest";
+import {BeatMetadata, BEATS_MANIFEST} from "../../../../assets/beats/beats-manifest";
 
 @Injectable({ providedIn: 'root' })
 export class BeatAdapter implements IManageBeats {
