+++ frontend/src/app/ui/services/sequencer/sequencer.service.ts
@@ -8,7 +8,7 @@ import { SequencerState } from "src/types/engine";
 import { SequencerViewModel } from "../../components/sequencer/sequencer.viewmodel";
 import { Effect, Option } from "effect";
 import { Track } from "src/app/domain/track";
-import { BEATS_MANIFEST, BeatMetadata } from './beats-manifest';
+import { BEATS_MANIFEST, BeatMetadata } from '../../../../assets/beats/beats-manifest';
 
 @Injectable({ providedIn: 'root' })
 export class SequencerService {
