+++ frontend/src/assets/beats/beats-manifest.ts
